Move the Ollama code for mmodelsyncer to internal/modelsyncer/ollama
'intenral/ollama/' will have small helper librariers.
