+++ engine/internal/modelsyncer/ollama/env.go
